pyproject.toml
[project]
name = "griid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "moderngl>=5.12.0",
    "moderngl-window>=3.1.1",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pyfluidsynth>=1.3.4",
    "pygame>=2.6.1",
    "pyyaml>=6.0.2",
    "rtmidi>=2.5.0",
    "sounddevice>=0.5.1",
    "soundfile>=0.13.1",
    "watchdog>=6.0.0",
]
